\name{armatblv-package}
\alias{armatblv-package}
\alias{armatblv}
\docType{package}
\title{
\packageTitle{armatblv}
}
\description{
\packageDescription{armatblv}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{armatblv}

Maintainer: \packageMaintainer{armatblv}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = armatblv)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{armatblv}}
%\section{Documentation Index}{\packageIndices{armatblv}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
