% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSE.R
\name{LSE}
\alias{LSE}
\title{Log of the sum of exponentials}
\usage{
LSE(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric value with log(sum(exp(x)))
}
\description{
Log of the sum of exponentials
}
\examples{
LSE(c(1, 10, 2, 4)))
}
