% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_mlog_like.R
\name{mc_mlog_like}
\alias{mc_mlog_like}
\title{Monte Carlo approximation for the marginal likelihood}
\usage{
mc_mlog_like(N, x, alpha, beta, kappa, phi, sigma, E_post_sample_list)
}
\arguments{
\item{N}{integer value with the Monte Carlo sample size}

\item{x}{numeric matrix with the data}

\item{alpha}{numeric matrix with estimated alpha}

\item{beta}{numeric matrix (number of items x 1) with estimated beta}

\item{kappa}{numeric matrix (number of items x 1) with estimated kappa}

\item{phi}{numeric value with estimated phi}

\item{sigma}{numeric value with estimated sigma}

\item{E_post_sample_list}{List where each element is posterior sample of E (iter x Latent dimension)}
}
\value{
log of the monte carlo sample for the likelihood
}
\description{
Monte Carlo approximation for the marginal likelihood
}
\examples{
#
}
